2007-10-26  Dan Williams  <dcbw@redhat.com>
	Patch from Helmut Schaa <hschaa@suse.de> (and more bits from me)

	* src/NetworkManagerAP.c
	  src/NetworkManagerAP.h
	  libnm-glib/nm-access-point.c
	  libnm-glib/nm-access-point.h
		- Make 'rate' property a guint32 to better match with WEXT and
			wpa_supplicant and to allow representation of higher bitrates

	* src/nm-device-802-11-wireless.c
	  introspection/nm-device-802-11-wireless.xml
	  libnm-glib/nm-device-802-11-wireless.c
	  libnm-glib/nm-device-802-11-wireless.h
		- Make 'bitrate' property a guint32 to match AP 'rate' property type

	* src/nm-device-802-3-ethernet.c
	  src/nm-device-802-3-ethernet.h
	  introspection/nm-device-802-3-ethernet.xml
	  libnm-glib/nm-device-802-3-ethernet.c
	  libnm-glib/nm-device-802-3-ethernet.h
		- Make 'speed' property a guint32 to match other speed/rate types
		- Make nm_device_802_3_ethernet_get_speed() static

	* test/nm-tool.c
		- Update for the changes above



git-svn-id: http://svn-archive.gnome.org/svn/NetworkManager/trunk@3030 4912f4e0-d625-0410-9fb7-b9a5a253dbdc

+++ src/NetworkManagerAP.c
@@ -43,7 +43,7 @@ typedef struct
 	int				mode;		/* from IW_MODE_* in wireless.h */
 	gint8			strength;
 	guint32			freq;		/* Frequency in GHz * 1000; ie 2.412 == 2412 */
-	guint16			rate;
+	guint32			rate;
 
 	guint32			flags;		/* General flags */
 	guint32			wpa_flags;	/* WPA-related flags */
@@ -965,9 +965,9 @@ nm_ap_set_freq (NMAccessPoint *ap,
  * Get/set functions for rate
  *
  */
-guint16 nm_ap_get_rate (NMAccessPoint *ap)
+guint32 nm_ap_get_rate (NMAccessPoint *ap)
 {
-	guint16 rate;
+	guint32 rate;
 
 	g_return_val_if_fail (NM_IS_AP (ap), 0);
 
@@ -977,7 +977,7 @@ guint16 nm_ap_get_rate (NMAccessPoint *ap)
 }
 
 void
-nm_ap_set_rate (NMAccessPoint *ap, guint16 rate)
+nm_ap_set_rate (NMAccessPoint *ap, guint32 rate)
 {
 	NMAccessPointPrivate *priv;
