+++ src/NetworkManagerAP.h
@@ -93,8 +93,8 @@ void				nm_ap_set_strength		(NMAccessPoint *ap, gint8 strength);
 guint32			nm_ap_get_freq			(NMAccessPoint *ap);
 void				nm_ap_set_freq			(NMAccessPoint *ap, guint32 freq);
 
-guint16			nm_ap_get_rate			(NMAccessPoint *ap);
-void				nm_ap_set_rate			(NMAccessPoint *ap, guint16 rate);
+guint32			nm_ap_get_rate			(NMAccessPoint *ap);
+void				nm_ap_set_rate			(NMAccessPoint *ap, guint32 rate);
 
 gboolean			nm_ap_get_invalid		(const NMAccessPoint *ap);
 void				nm_ap_set_invalid		(NMAccessPoint *ap, gboolean invalid);
